refactor: split UserTable into server/client, add provider/key columns
